[fix] クラスの相互参照やヘッダ読み込みの順番によるエラーの修正
