+++ Lonely/Lonely/Game/Scene/TitleScene/Cursor.cpp
@@ -1,0 +1,15 @@
+/**
+* @file Cursor.cpp
+* @brief Cursorクラスのソースファイル
+* @author 佐川　嗣苑
+*/
+
+#include "Cursor.h"
+
+Cursor::Cursor()
+{
+}
+
+Cursor::~Cursor()
+{
+}
