+++ Lonely/Lonely/Game/Scene/TitleScene/Cursor.h
@@ -1,21 +1,24 @@
+/**
+* @file Cursor.h
+* @brief Cursorクラスのヘッダファイル
+* @author 佐川　嗣苑
+*/
+
 #pragma once
 
 #include "../../../GameLib/ObjectManager/Object.h"
 
+/**
+* @brief タイトルメニュのカーソル関係のクラス
+*/
 class Cursor : public Object
 {
 public:
 	Cursor();
 	~Cursor();
 
+
+
 private:
 
 };
-
-Cursor::Cursor()
-{
-}
-
-Cursor::~Cursor()
-{
-}
