+++ Lonely/Lonely/Game/Scene/TitleScene/TitleLogo.cpp
@@ -1,0 +1,15 @@
+/**
+* @file TitleLogo.cpp
+* @brief TitleLogoクラスのソースファイル
+* @author 佐川　嗣苑
+*/
+
+#include "TitleLogo.h"
+
+TitleLogo::TitleLogo()
+{
+}
+
+TitleLogo::~TitleLogo()
+{
+}
