+++ Lonely/Lonely/Game/Scene/TitleScene/TitleLogo.h
@@ -1,7 +1,16 @@
+/**
+* @file TitleLogo.h
+* @brief TitleLogoクラスのヘッダファイル
+* @author 佐川　嗣苑
+*/
+
 #pragma once
 
 #include "../../../GameLib/ObjectManager/Object.h"
 
+/**
+* @brief タイトルロゴの処理をまとめたクラス
+*/
 class TitleLogo : public Object
 {
 public:
@@ -11,11 +20,3 @@ public:
 private:
 
 };
-
-TitleLogo::TitleLogo()
-{
-}
-
-TitleLogo::~TitleLogo()
-{
-}
