+++ Lonely/Lonely/Game/Scene/TitleScene/TitleMenu1.cpp
@@ -3,3 +3,14 @@
 * @brief TitleMenu1クラスのソースファイル
 * @author 佐川　嗣苑
 */
+
+#include "TitleMenu1.h"
+
+
+TitleMenu1::TitleMenu1()
+{
+}
+
+TitleMenu1::~TitleMenu1()
+{
+}
