+++ Lonely/Lonely/Game/Scene/TitleScene/TitleMenu2.cpp
@@ -3,3 +3,13 @@
 * @brief TitleMenu2クラスのソースファイル
 * @author 佐川　嗣苑
 */
+
+#include "TitleMenu2.h"
+
+TitleMenu2::TitleMenu2()
+{
+}
+
+TitleMenu2::~TitleMenu2()
+{
+}
