+++ Lonely/Lonely/GameLib/ObjectManager/Object.cpp
@@ -5,3 +5,30 @@
 */
 
 #include "Object.h"
+
+
+Object::Object()
+{
+}
+
+Object::~Object()
+{
+}
+
+bool Object::Initialize()
+{
+	return false;
+}
+
+void Object::Finalize()
+{
+}
+
+void Object::Update()
+{
+}
+
+void Object::Render()
+{
+}
+
